% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr.R
\name{clr}
\alias{clr}
\title{clr}
\usage{
clr(x)
}
\arguments{
\item{x}{dataset}
}
\description{
Source: https://cran.r-project.org/web/packages/robCompositions/index.html
}
\examples{
data(iris)
clr(iris[,1:4])
biplot(clr(iris[,1:4]))
}
\keyword{accoring}
\keyword{clr}
\keyword{robCompositions}
\keyword{the}
\keyword{to}
\keyword{tranformation}
