% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doclrcol.R
\name{doclrcol}
\alias{doclrcol}
\title{doclrcol}
\usage{
doclrcol(a, zero = T, na = T, base = 10)
}
\arguments{
\item{a}{for dim}

\item{zero}{zero=T}

\item{na}{na=T}

\item{base}{The logarithm base is set to 10}
}
\description{
This function allows you to process clr transformation in columns.
}
\examples{
doclrcol(gpr**2)
}
\keyword{clr}
