% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpr.R
\docType{data}
\name{gpr}
\alias{gpr}
\title{Data from gpr geophysical measurement.}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\source{
\href{https://en.wikipedia.org/wiki/Ground-penetrating_radar}
}
\usage{
data(gpr)
}
\description{
Data from gpr geophysical measurement.
Data from gpr geophysical measurement.
Data from gpr geophysical measurement.
Data from gpr geophysical measurement.
}
\examples{
data(gpr)
}
\references{

}
\keyword{datasets}
