% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integerxy.R
\name{integerxy}
\alias{integerxy}
\title{integerxy}
\usage{
integerxy(a, x = 1, y = 2)
}
\arguments{
\item{a}{a = dataframe, x = position in matrix, y = position in matrix}
}
\description{
integerxy
}
\examples{
integerxy(iris[,1:3])
}
\keyword{in}
\keyword{integer}
\keyword{matrix}
\keyword{position}
\keyword{to}
