% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doclrrow.R
\name{doclrrow}
\alias{doclrrow}
\title{doclrrow}
\usage{
doclrrow(a, zero = T, na = T, base = 10)
}
\arguments{
\item{a}{for dim}

\item{zero}{zero=T}

\item{na}{na=T}

\item{base}{The logarithm base is set to 10}
}
\description{
This function allows you to process clr transformation in rows.
}
\examples{
doclrrow(gpr**2)
}
\keyword{clr}
