% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpr.R
\docType{data}
\name{gpr}
\alias{gpr}
\title{Data from gpr geophysical measurement.}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\source{
\href{https://phenome.jax.org/projects/Moore1b}{QTL Archive}
}
\usage{
data(gpr)
}
\description{
Data from gpr geophysical measurement.
Data from gpr geophysical measurement.
Data from gpr geophysical measurement.
Data from gpr geophysical measurement.
}
\examples{
data(gpr)
}
\references{
Moore et al. (2013) Genetics 195:1077-1086
(\href{https://www.ncbi.nlm.nih.gov/pubmed/23979570}{PubMed})
}
\keyword{datasets}
